TASK 709: Run the documented strict Clippy contract in every CI gate
CI's workspace Clippy omitted --features rust-tests, so every test-only target
and every rust-tests-gated path was linted in its disabled form and a warning
there could not fail the gate. The command now matches the documented contract.

Turning it on exposed 33 warnings that -D warnings would have turned into a red
CI, so shipping the tightening alone would have shipped a broken gate. They are
cleared: 17 through cargo clippy --fix, the rest by hand.

Two were real defects. WintunCleanupState and its methods were defined outside
the target_os = "windows" gate that contains every one of their constructors
and consumers, so they compiled on every other platform as an unreachable type;
the definition, impl, and test now sit under that gate.
MemoryPool::accounting_snapshot was gated on rust-tests while every consumer is
an in-crate cfg(test) module, so it was dead in rust-tests library builds. An
unused duplicate packet-number encoder was removed from the header test.

The Clippy matrix grew from 8 to 21 feature entries, and a new
feature-matrix-coverage job fails when Cargo.toml declares a feature the matrix
omits, and equally when the matrix lists one Cargo.toml does not declare. It
found test-suite and experimental missing on its first run. tun-windows and
tun-ios are excluded with a stated platform reason.

Tauri host Clippy with -D warnings already existed; that part of the finding
was stale and was verified rather than duplicated.

Adjacent finding filed as TODO-875 rather than absorbed: --all-targets reports
2619 passed, 2 failed, both in the binary target and both reproducing on a
clean tree, invisible to the --lib baseline.

+++ src/memory_lock.rs
@@ -410,7 +410,7 @@ fn mlockall_flags_for_limit(current_limit: libc::rlim_t) -> libc::c_int {
     mlockall_flags_for_budget(if current_limit == libc::RLIM_INFINITY {
         MemoryLockLimit::Unlimited
     } else {
-        MemoryLockLimit::Finite(current_limit as u64)
+        MemoryLockLimit::Finite(current_limit)
     })
 }
 
@@ -459,7 +459,7 @@ fn classify_memlock_limit(
         Ok(current_limit) if current_limit == libc::RLIM_INFINITY => {
             Ok((MemoryLockLimit::Unlimited, false))
         }
-        Ok(current_limit) => Ok((MemoryLockLimit::Finite(current_limit as u64), false)),
+        Ok(current_limit) => Ok((MemoryLockLimit::Finite(current_limit), false)),
         Err(error) if failure_policy == MemoryLockFailurePolicy::BestEffort => {
             log::warn!("RLIMIT_MEMLOCK query failed: {error}; using MCL_CURRENT fallback");
             Ok((MemoryLockLimit::Unknown, true))
